Revert "Sohail commit upto sharing"

+++ Shopisthan-Frontend/src/actions/auth.action.js
@@ -1,6 +1,6 @@
 import axiosIntance from "../helpers/axios";
 import axios from "../helpers/axios";
-import { authConstants, cartConstants } from "./constants";
+import { authConstants } from "./constants";
 
 export const login = (user) => {
 
@@ -91,21 +91,24 @@ export const isUserLoggedIn = () => {
 
 
 export const signout = () => {
-  return async (dispatch) => {
-    dispatch({ type: authConstants.LOGOUT_REQUEST });
-    // localStorage.removeItem('user');
-    // localStorage.removeItem('token');
-    localStorage.clear();
-    dispatch({ type: authConstants.LOGOUT_SUCCESS });
-    dispatch({ type: cartConstants.RESET_CART });
-    //const res = await axios.post(`/admin/signout`);
-    // if(res.status === 200){
+  return async dispatch => {
 
-    // }else{
-    //     dispatch({
-    //         type: authConstants.LOGOUT_FAILURE,
-    //         payload: { error: res.data.error }
-    //     });
-    // }
-  };
-};
+    dispatch({ type: authConstants.LOGOUT_REQUEST });
+
+    const res = await axios.post(`/admin/signout`)
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS
+      });
+
+    } else {
+
+      dispatch({
+        type: authConstants.LOGIN_FAILURE,
+        payload: { error: res.data.error }
+      });
+    }
+
+  }
+}
